refactor: migrate noteService to neverthrow
