+++ backend/src/api/routes/notes.ts
@@ -1,7 +1,7 @@
-import { NoteService } from '../../services/note.service';
+import type { ResultAsync } from 'neverthrow';
+import { createNoteService } from '../../services/note.service';
 import { ThreadService } from '../../services/thread.service';
-import { DeleteService } from '../../services/delete.service';
-import { db } from '../../db';
+import { errorToStatusCode, type NoteError } from '../../errors/domain-errors';
 import {
   validateCreateNote,
   validateUpdateNote,
@@ -11,75 +11,104 @@ import {
 import { requireAuth } from '../../middleware/auth.middleware';
 import type { NoteListResponse, NoteDetailResponse, ErrorResponse } from '@thread-note/shared/types';
 import { serialize } from '../../types/api';
-import { Hono } from 'hono';
-import { errorToStatusCode, type NoteError } from '../../errors/domain-errors';
+import { db } from '../../db';
+import { createRouter } from './router';
 
-/** Convert NoteError to ErrorResponse */
+/**
+ * NoteError を ErrorResponse に変換
+ */
 const toErrorResponse = (error: NoteError): ErrorResponse => ({
   error: error._tag,
   message: error.message,
 });
 
-const noteService = new NoteService({ db });
-const threadService = new ThreadService({ db });
-const deleteService = new DeleteService({ db });
+/**
+ * Result のマッチングによるレスポンス生成
+ *
+ * この関数により、ルートハンドラ内で一貫したエラーハンドリングが可能
+ */
+const handleResult = async <T>(
+  c: { json: (body: unknown, status?: number) => Response; body: (body: unknown, status?: number) => Response },
+  resultAsync: ResultAsync<T, NoteError>,
+  onSuccess: (value: T) => { body: unknown; status?: number } | Promise<{ body: unknown; status?: number }>
+) => {
+  const result = await resultAsync;
 
-const app = new Hono()
+  if (result.isOk()) {
+    const response = await onSuccess(result.value);
+    // 204 No Content の場合は body メソッドを使用
+    if (response.status === 204) {
+      return c.body(response.body, response.status);
+    }
+    return c.json(response.body, response.status);
+  }
+
+  const error = result.error;
+  return c.json(toErrorResponse(error), errorToStatusCode(error));
+};
+
+const functionalNoteService = createNoteService({ db });
+
+const app = createRouter()
   // GET /api/notes - List root notes
   .get('/', requireAuth, validatePagination, async (c) => {
     const { limit, offset } = c.req.valid('query');
-    const result = await noteService.getRootNotes(limit, offset);
 
-    const response: NoteListResponse = {
-      notes: result.notes.map(serialize),
-      total: result.total,
-      hasMore: result.hasMore,
-    };
-
-    return c.json(response);
+    return handleResult(c, functionalNoteService.getRootNotes(limit, offset), (data) => {
+      const response: NoteListResponse = {
+        notes: data.notes.map(serialize),
+        total: data.total,
+        hasMore: data.hasMore,
+      };
+      return { body: response };
+    });
   })
+
   // POST /api/notes - Create note
-  .post('/', validateCreateNote, async (c) => {
+  .post('/', requireAuth, validateCreateNote, async (c) => {
     const data = c.req.valid('json');
-    const note = await noteService.createNote(data);
-    return c.json(serialize(note), 201);
+
+    return handleResult(c, functionalNoteService.createNote(data), (note) => ({
+      body: serialize(note),
+      status: 201,
+    }));
   })
+
   // GET /api/notes/:id - Get note with thread
-  // Uses neverthrow Result type for type-safe error handling
-  .get('/:id', validateNoteId, async (c) => {
+  .get('/:id', requireAuth, validateNoteId, async (c) => {
     const { id } = c.req.valid('param');
     const includeThread = c.req.query('includeThread') !== 'false';
 
-    const result = await noteService.getNoteByIdResult(id);
+    return handleResult(c, functionalNoteService.getNoteById(id), async (note) => {
+      const threadService = new ThreadService({ db });
+      const thread = includeThread ? await threadService.getThread(id) : [];
 
-    if (result.isErr()) {
-      const error = result.error;
-      return c.json(toErrorResponse(error), errorToStatusCode(error));
-    }
-
-    const note = result.value;
-    const thread = includeThread ? await threadService.getThread(id) : [];
-
-    const response: NoteDetailResponse = {
-      note: serialize(note),
-      thread: thread.map(serialize),
-    };
-
-    return c.json(response);
+      const response: NoteDetailResponse = {
+        note: serialize(note),
+        thread: thread.map(serialize),
+      };
+      return { body: response };
+    });
   })
+
   // PUT /api/notes/:id - Update note
   .put('/:id', requireAuth, validateNoteId, validateUpdateNote, async (c) => {
     const { id } = c.req.valid('param');
     const data = c.req.valid('json');
 
-    const updated = await noteService.updateNote(id, data);
-    return c.json(serialize(updated));
+    return handleResult(c, functionalNoteService.updateNote(id, data), (note) => ({
+      body: serialize(note),
+    }));
   })
+
   // DELETE /api/notes/:id - Delete note (cascade)
   .delete('/:id', requireAuth, validateNoteId, async (c) => {
     const { id } = c.req.valid('param');
-    await deleteService.deleteNote(id);
-    return c.body(null, 204);
+
+    return handleResult(c, functionalNoteService.deleteNote(id), () => ({
+      body: null,
+      status: 204,
+    }));
   });
 
 export default app;
